// src/app/shared/generated/model/agentMcpToolRule.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ToolPermission } from './toolPermission';


export interface AgentMcpToolRule { 
    id?: string;
    modificationCount?: number;
    toolName?: string;
    toolDescription?: string;
    allowed?: ToolPermission;
}